Refactoring: use same component for multicolumn posts and thread posts, add flow types
