+++ assembl/static2/tests/unit/components/debate/common/post/post.spec.jsx
@@ -1,18 +1,18 @@
 import React from 'react';
 import ShallowRenderer from 'react-test-renderer/shallow';
 
-import { PostFolded } from '../../../../../js/app/components/debate/thread/post';
+import FoldedPost from '../../../../../../js/app/components/debate/common/post/foldedPost';
 
-describe('PostFolded component', () => {
+describe('FoldedPost component', () => {
   it('should render a folded post', () => {
     const renderer = new ShallowRenderer();
-    renderer.render(<PostFolded nbPosts={1} />);
+    renderer.render(<FoldedPost nbPosts={1} />);
     const result = renderer.getRenderOutput();
     expect(result).toMatchSnapshot();
   });
   it('should render several folded posts', () => {
     const renderer = new ShallowRenderer();
-    renderer.render(<PostFolded nbPosts={3} />);
+    renderer.render(<FoldedPost nbPosts={3} />);
     const result = renderer.getRenderOutput();
     expect(result).toMatchSnapshot();
   });
